backend/requirements.txt
pandas
scikit-learn
joblib
PyMuPDF
lightgbm
google-generativeai
chromadb
flask
python-dotenv
google-generativeai
werkzeug
uvicorn[standard]
asgiref
sentence-transformers
